Create missing_number.py
Missing Number
Given an array nums containing N distinct numbers in the range [1, N + 1], return the only number in the range that is missing from the array.

Input
First line of the input file consist of t which is the number of test cases. Then t test cases follow. Each test case consists of two lines. First line contains N which is the size of the array. Next line consists of N space separated integers which is the content of the array.

Output
For each test case, output the missing number in a separate line.

Example
Input:
2
5
1 3 2 5 6
3
2 4 3

Output:
4
1

Explanation In the above example, for the first test case the array is supposed to contain numbers from 1 to 6 (5 + 1). It is missing 4, so 4 is the answer.
In the second test case. the array is supposed to contain numbers from 1 to 4. It is missing 1, so 1 is the answer.