Create two_sum.py
Two Sum
Given an array (arr[]) of integers and integer S and the task is to find the indices of the two element that sum up to a S. If the indices is not found then print "-1 -1". If there are multiple possible answers, print the answer with the lowest first index and lowest second index.

Input
Each input file starts with the number of test cases t. After that t test cases follow :
Each test case consists of two lines :

First Line contains N which is the size of array and S which is the expected sum.
Second line consists of N space separated integers of the array.
Output
For each test case output a single line consisting of the two indices which sum up to S.

Example
Input:
3
5 20
10 21 2 20 10
3 20
10 10 10
3 25
10 10 10

Output:
0 4
0 1
-1 -1